fix: переименовал файл activivty_componeit.html на activivty_component.htm.     fix: добавил в файл migrations/ в gitignore     fix: создал файл theme_settings.py и импортировал в settings.py
