+++ .gitignore
@@ -24,4 +24,4 @@ media/files/Telegram.exe
 
 cred/
 
-
+migrations/
