+++ chat/templates/base/activivty_component.html
