+++ chat/templates/base/home.html
@@ -18,7 +18,7 @@
     <hr class="divider" />
     <h3 class="recent-header">Recent correspondence</h3>
     <!-- Сообщения -->
-    <div class="activity">{% include "base/activivty_componeit.html" %}</div>
+    <div class="activity">{% include "base/activivty_component.html" %}</div>
     <!-- КОНЕЦ Сообщения -->
   </div>
 </div>
